Factored out the majority of the HTMLContentPlugin code into a new HTMLTemplateUtility class, so the code can also be used by the forthcoming '<brjs-templates>' tag.
